Moved the sql.Open() code into stored_requests/
